refactor: extract AppState issue-mutation extension (#579, #430)
Seventh slice of #430 (AppState decomposition). Extracts 3 issue-mutation
do-catch delegators (updateExtractedIssue, setIssueSelection,
setAllIssuesSelected) into a new
Sources/BugNarrator/AppState+IssueMutation.swift as
`extension AppState { ... }`.

Range: AppState.swift lines 1078-1100 (23 lines, 862 bytes, SHA-256
7d92ddbd92cf2f5d92868b60cfb0a48dde531bb8527364205a9d0b43a600a20d).
All 3 methods use the byte-identical do-catch shape:

    do { try issueExtractionController.<method>(...) }
    catch { issueMutationFailurePresenter.presentFailure(error) }

Same shape as #576's session-library cluster (3x sessionLibrary do-catch
delegators). Combined with this slice, the codebase now has 6 identical
do-catch delegators across 2 different presenters — the rule-of-three
threshold for a shared helper is met. Filed as #578 for follow-up;
timing plan comment on #578 says to schedule after this slice.

Full triple-proof stack applied per the established protocol:
- Endpoint verification at ticket-draft time AND at build time (line
  1078 = func header, 1100 = closing }, 1101 = trailing blank)
- SHA gate matched at both spec-review (against origin) and post-extract
  (against new-file body)
- Byte-diff against origin/main is silent
- Delegator-vs-orchestrator test: all 3 pass

Extension inherits @MainActor from AppState per SE-0316 — not
re-annotated. New file imports `import Foundation` only (UUID +
same-module ExtractedIssue). No dead-import cleanup — AppState.swift
still needs Combine (ObservableObject/@Published) and Foundation
(UUID/URL/TimeInterval).

662 unit tests pass unchanged; docs-sync validation passes.

Pre-build spec reviewed by codex in a fresh session (0 findings — clean
pass, upstream ticket-drafting discipline continues to make codex's
spec-review nit catches redundant).
